Add files via upload
The idea would be that those who are in the flood area must evacuate the flood area and go to one of the two safe zones (bell centre and olympic stadium - in green in the map above). The people in the flooded area would have to walk to the nearest stop outside of the flooded area. The busses would have to include the closest safe zone on the route for those who need to evacuate.